feat(launch file refactoring): moved common_functions_launch file into launch folder
